extract to ExpressionMappingBodyBuilder
