+++ src/Riok.Mapperly/Descriptors/MappingBodyBuilders/MappingBodyBuilder.cs
@@ -41,11 +41,11 @@ public class MappingBodyBuilder(MappingCollection mappings)
                 case UserDefinedNewInstanceMethodMapping mapping:
                     UserMethodMappingBodyBuilder.BuildMappingBody(ctx, mapping);
                     break;
-                case UserDefinedExpressionMethodMapping mapping:
-                    UserMethodMappingBodyBuilder.BuildMappingBody(ctx, mapping);
-                    break;
                 case UserDefinedExistingTargetMethodMapping mapping:
                     UserMethodMappingBodyBuilder.BuildMappingBody(ctx, mapping);
+                    break;
+                case UserDefinedExpressionMethodMapping mapping:
+                    ExpressionMappingBodyBuilder.BuildMappingBody(ctx, mapping);
                     break;
                 case UserDefinedNewInstanceRuntimeTargetTypeParameterMapping mapping:
                     RuntimeTargetTypeMappingBodyBuilder.BuildMappingBody(ctx, mapping);
